pyproject.toml
[project]
name = "miltrack"
version = "0.1.0"
description = "Live military aircraft tracker with conflict event overlay"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "httpx>=0.28.0",
    "websockets>=16.0",
    "python-dotenv>=1.2.1",
    "databricks-sdk>=0.95.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]
